+++ dll/ime/msctfime/CMakeLists.txt
@@ -1,12 +1,8 @@
 
-include_directories(
-    ${REACTOS_SOURCE_DIR}/win32ss/include)
-
 spec2def(msctfime.ime msctfime.spec)
 
 list(APPEND SOURCE
     bridge.cpp
-    functions.cpp
     inputcontext.cpp
     misc.cpp
     msctfime.cpp
