+++ dll/ime/msctfime/msctfime.h
@@ -35,18 +35,10 @@
 
 #include <wine/debug.h>
 
-EXTERN_C BOOLEAN WINAPI DllShutdownInProgress(VOID);
+extern CRITICAL_SECTION g_csLock;
 
-HRESULT InitDisplayAttrbuteLib(PCIC_LIBTHREAD pLibThread);
-HRESULT UninitDisplayAttrbuteLib(PCIC_LIBTHREAD pLibThread);
-
-static inline HIMC GetActiveContext(VOID)
-{
-    HWND hwndFocus = ::GetFocus();
-    if (!hwndFocus)
-        hwndFocus = ::GetActiveWindow();
-    return ::ImmGetContext(hwndFocus);
-}
+typedef CicArray<GUID> CDispAttrPropCache;
+extern CDispAttrPropCache *g_pPropCache;
 
 DEFINE_GUID(GUID_COMPARTMENT_CTFIME_DIMFLAGS,        0xA94C5FD2, 0xC471, 0x4031, 0x95, 0x46, 0x70, 0x9C, 0x17, 0x30, 0x0C, 0xB9);
 DEFINE_GUID(GUID_COMPARTMENT_CTFIME_CICINPUTCONTEXT, 0x85A688F7, 0x6DC8, 0x4F17, 0xA8, 0x3A, 0xB1, 0x1C, 0x09, 0xCD, 0xD7, 0xBF);
@@ -55,13 +47,11 @@ DEFINE_GUID(GUID_MODEBIAS_NUMERIC,                   0x4021766C, 0xE872, 0x48FD,
 DEFINE_GUID(GUID_MODEBIAS_URLHISTORY,                0x8B0E54D9, 0x63F2, 0x4C68, 0x84, 0xD4, 0x79, 0xAE, 0xE7, 0xA5, 0x9F, 0x09);
 DEFINE_GUID(GUID_MODEBIAS_DEFAULT,                   0xF3DA8BD4, 0x0786, 0x49C2, 0x8C, 0x09, 0x68, 0x39, 0xD8, 0xB8, 0x4F, 0x58);
 DEFINE_GUID(GUID_PROP_MODEBIAS,                      0x372E0716, 0x974F, 0x40AC, 0xA0, 0x88, 0x08, 0xCD, 0xC9, 0x2E, 0xBF, 0xBC);
-
 #define GUID_MODEBIAS_NONE GUID_NULL
 
 #include "resource.h"
 
 #include "bridge.h"
-#include "functions.h"
 #include "inputcontext.h"
 #include "misc.h"
 #include "profile.h"
